Реализован поиск пути

+++ metro/src/main/java/com/metropaths/Path.java
@@ -3,10 +3,10 @@ package com.metropaths;
 import java.util.List;
 
 public class Path {
-    public int lessTime;
+    public Double lessTime;
     public List<Integer> path;
 
-    public Path (Integer lessTime, List<Integer> path) {
+    public Path (Double lessTime, List<Integer> path) {
         this.lessTime = lessTime;
         this.path = path;
     }
